Restructure config
