+++ ecosystem/nft-metadata-crawler-parser/src/lib.rs
@@ -1,5 +1,6 @@
 // Copyright © Aptos Foundation
 
+pub mod config;
 pub mod models;
 pub mod schema;
 pub mod utils;
